requirements.txt
MySQL-python==1.2.5
pandas==1.5.1
python-dotenv==0.21.0
SQLAlchemy==1.4.44
